commit 07.10.2020  19:16

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoJDBCImpl.java
@@ -80,7 +80,7 @@ public class UserDaoJDBCImpl implements UserDao {
                 user.setId(resultSet.getLong(1));
                 user.setName(resultSet.getString(2));
                 user.setLastName(resultSet.getString(3));
-                user.setAge((byte) resultSet.getLong(4));
+                user.setAge (resultSet.getByte(4));
                 allUsers.add(user);
             }
             stmt.executeUpdate(sql);
